Halve corvettes and frigates
A Terran frigate measured 6.34 units in a fight that happens inside 200 and
under weapons that reach 300, so the smallest thing anyone flies was a third
the size of the engagement it was in. `RUNG.frigate` is 3.5/64 now: the frigate
is 3.17 and the corvette 1.59, across all four navies. A rung is a cell size
and the ladder is a consequence of it, so this is one number rather than eight
tables.

Destroyers and heavy cruisers are untouched, which is the owner's call and
costs the ladder: a destroyer is 2.95 times a frigate rather than 1.5, so the
big step is now frigate to destroyer and the small one corvette to frigate.

The civil lighter gets its own rung at the old cell. It shared the frigate's
and is neither a frigate nor a corvette, so halving it would have shrunk a
cargo tender because a warship two tables away changed size.

**Two things followed and neither is cosmetic.**

Plate mass and hull go as the CUBE of the cell, so a Terran frigate carries 121
hull points where it carried 300 and a beam still does 27.5. These classes are
markedly more fragile: the playthrough now reaches a victory in six turns with
both sides down to one ship by turn five. `tests/turn.rs` asks its damage model
questions of heavy cruisers, because a frigate no longer survives being pounded
on a single volume long enough for that volume to die, and the model is the
same at every rung.

`FRIGATE_CELL` in `data.rs` is a REFERENCE cell, not a statement about
frigates. `PLATE_UM`, `HULL_MILLI` and `MOUNT_RADIUS` are authored per cell of
that size and every rung scales by its own `rung_cell` over it, so halving it
alongside `RUNG.frigate` puts the ratio back to one and silently undoes the
change: the core and the editor then disagreed about a frigate's mass by 80
percent. Its comment says so now.

The catch radius needed no edit for the same reason, and `tests/mounts.rs` did:
it read the CONSTANT rather than `mount_radius(class)`, so it reported a
frigate as having twice its actual catch radius the moment the rung moved. It
asks for the class's own now. `measure_fleet.mjs --fixtures` also emitted
`radius_cells: 22` for a radius that landed on a whole number, which Rust
refuses where it wants an f32: it goes through `f32` like every other float, so
the generated table compiles on the hulls whose numbers come out roundest.

Every gate: 29 prototype, 96 Rust, 80 web, 13 server, the class table matches
the fleet at 23 classes with every hull back at 84 to 85 percent of its berth,
and a match plays through to a victory.

+++ tools/measure_fleet.mjs
@@ -265,7 +265,11 @@ if (process.argv.includes('--sync') || process.argv.includes('--check')) {
     console.log(`        (ShipClassId::${ID[r.key] ?? r.key},`);
     console.log(`         &[${r.partList.join(', ')}],`);
     console.log(`         Geometry { plate_cells: ${r.geo.plateCells}, `
-      + `ext: [${r.geo.ext.join(', ')}], radius_cells: ${r.geo.radiusCells}, `
+      // Through `f32` like every other float. A radius that lands on a whole
+      // number prints as `22` in JavaScript and Rust refuses an integer where
+      // it wants an f32, so the generated table would not compile on exactly
+      // the hulls whose numbers came out roundest.
+      + `ext: [${r.geo.ext.join(', ')}], radius_cells: ${f32(r.geo.radiusCells)}, `
       + `fouled: ${r.geo.fouled} },`);
     console.log(`         Expect { mass: ${f32(st.mass)}, hull: ${f32(st.hull)}, `
       + `radius: ${f32(st.radius)}, accel_fwd: ${f32(st.accelFwd)}, `
